Permissions branch

+++ backend/forms/views.py
@@ -12,10 +12,20 @@ from rest_framework.views import APIView
 
 from forms.models import FormsTable
 from forms.serializers import FormsSerializer
-
+from users.utils import check_permissions, PERMISSIONS_PAGE_FROM_MANAGER, PERMISSIONS_PAGE_FROM_EDIT_EVENTS, PERMISSIONS_PAGE_FROM_WATCHING_EVENTS
 import time
 import os
 import csv
+
+def check_permissions(request,permissions_array):
+    token=request.headers['Authorization']
+    token=token.split(" ")
+    token=token[1]
+    permission=utils.check_token_not_login(token)
+    if permission in permissions_array:
+        return True
+    else:
+        return False
 
 @csrf_exempt
 def forms_list(request, event_type):
@@ -26,20 +36,22 @@ def forms_list(request, event_type):
               Return all rows from table when no url specified.
         DELETE - Delete all table content. 
     '''
-    
+    if not check_permissions(request,[PERMISSIONS_PAGE_FROM_MANAGER,PERMISSIONS_PAGE_FROM_EDIT_EVENTS,PERMISSIONS_PAGE_FROM_WATCHING_EVENTS]):
+        return HttpResponse(status=status.HTTP_401_FORBIDDEN)
     if request.method == 'GET':
-        if event_type == '':
-            forms = FormsTable.objects.all()
-        else:
-            forms = FormsTable.objects.filter(eventType=event_type)
-        
-        form_serializer = FormsSerializer(forms, many=True)
+            if event_type == '':
+                forms = FormsTable.objects.all()
+            else:
+                forms = FormsTable.objects.filter(eventType=event_type)
+            
+            form_serializer = FormsSerializer(forms, many=True)
 
-        return JsonResponse(form_serializer.data, safe=False)
+            return JsonResponse(form_serializer.data, safe=False) 
     
     elif request.method == 'DELETE':
         FormsTable.objects.all().delete()
         return HttpResponse(status=status.HTTP_204_NO_CONTENT)
+
 
 
 
@@ -49,10 +61,13 @@ def new_event_form(request):
         New Event Form load all its initial values from here.
         Relevant url: /api/event_forms/
     '''
+    if not check_permissions(request,[PERMISSIONS_PAGE_FROM_MANAGER, PERMISSIONS_PAGE_FROM_EDIT_EVENTS, PERMISSIONS_PAGE_FROM_WATCHING_EVENTS]):
+        return HttpResponse(status=status.HTTP_401_UNAUTHORIZED)
     if request.method == 'GET':
         dt = datetime.today()
-        payload = {'': '{}/{}/{}'.format(dt.day, dt.month, dt.year)}
+        payload = {'datetime': '{}/{}/{}'.format(dt.day, dt.month, dt.year)}
         return JsonResponse(payload, safe=False)
+            
 '''
 @csrf_exempt
 def existing_event_form(request, reference):
@@ -83,13 +98,15 @@ def download_file(request, path):
     '''
         Download file from given path in the request url.
     '''
+    if not check_permissions(request,PERMISSIONS_PAGE_FROM_MANAGER):
+        return HttpResponse(status=status.HTTP_401_UNAUTHORIZED)
     file_path = os.path.join(settings.MEDIA_ROOT, path)
     if os.path.exists(file_path):
         with open(file_path, 'rb') as fh:
             response = HttpResponse(fh.read(), content_type="application/vnd.ms-excel")
             response['Content-Disposition'] = 'inline; filename=' + os.path.basename(file_path)
             return response
-    return HttpResponse(status=status.HTTP_404_NOT_FOUND)
+    
 
 
 @csrf_exempt
@@ -121,14 +138,18 @@ class NewEventFrom(APIView):
     parser_classes = (MultiPartParser, FormParser)
 
     def post(self, request, *args, **kwargs):
+        if not check_permissions(request,[PERMISSIONS_PAGE_FROM_MANAGER,PERMISSIONS_PAGE_FROM_EDIT_EVENTS]):
+            return HttpResponse(status=status.HTTP_401_FORBIDDEN)
         form_serializer = FormsSerializer(data=request.data)
         if form_serializer.is_valid():
             form_serializer.save(reference = int(FormsTable.objects.aggregate(Max('reference'))['reference__max'] or 0) + 1)
             return JsonResponse(form_serializer.data, status=status.HTTP_201_CREATED ) 
         else:
-            return HttpResponse(form_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            return HttpResponse(status=status.HTTP_400_BAD_REQUEST)
 
     def put(self, request, reference,*args, **kwargs):
+        if not check_permissions(request,[PERMISSIONS_PAGE_FROM_MANAGER,PERMISSIONS_PAGE_FROM_EDIT_EVENTS]):
+            return HttpResponse(status=status.HTTP_401_FORBIDDEN)
         try: 
             event_form = FormsTable.objects.get(reference=reference) 
         except FormsTable.DoesNotExist: 
@@ -151,6 +172,8 @@ class NewEventFrom(APIView):
         return JsonResponse(form_serializer.data)
 
     def delete(self, request, reference, *args, **kwargs):
+        if not utils.check_permissions(request,[PERMISSIONS_PAGE_FROM_MANAGER]):
+            return HttpResponse(status=status.HTTP_401_FORBIDDEN)
         try: 
             event_form = FormsTable.objects.get(reference=reference)
         except FormsTable.DoesNotExist:
